changement dans RecetteRepository, on change le type de l'id

+++ Craft2Gether/src/main/java/fr/crafttogether/repositories/RecetteRepository.java
@@ -3,5 +3,5 @@ package fr.crafttogether.repositories;
 import fr.crafttogether.models.Recette;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecetteRepository extends JpaRepository<Recette, Long>  {
+public interface RecetteRepository extends JpaRepository<Recette, Integer>  {
 }
